requirements.txt
asgiref==3.4.1
Django==3.2.5
django-tempus-dominus==5.1.2.14
django-widget-tweaks==1.4.8
listview==0.1.0
mysql-client==0.0.1
numpy==1.21.2
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
